+++ src/app/BreakoutApp.hpp
@@ -2,7 +2,7 @@
 
 #include "App.hpp"
 #include "BreakoutLogic.hpp"
-#include "../detail/TextSurfaceCache.hpp"
+#include "../detail/TextTextureCache.hpp"
 #include <SDL2/SDL.h>
 #include <SDL2/SDL_ttf.h>
 
@@ -36,7 +36,7 @@ private:
     int hudHeight() const;
 
     TTF_Font* m_font = nullptr;
-    mutable monolith::detail::TextSurfaceCache m_textSurfaceCache;
+    mutable monolith::detail::TextTextureCache m_textTextureCache;
     monolith::breakout::Game m_game;
     bool m_paused = false;
     bool m_holdLeft = false;